MIT app inventor
This is our application for the project. The mobile application provides a user-friendly interface, offering real-time information on the current number of fish, inventory levels (food, acid, base), and graphical representations of pH and temperature trends over time. To collect real-time data, we employ an RTC module integrated with the Arduino. Users can input parameters such as the number of fish added or removed, average body weight, fish species, and inventory status through the app interface. This information is seamlessly communicated to the Arduino for system control.

In the event of critically low levels of food, base, or acid, the system promptly sends an alert to the mobile application, notifying the user to take necessary action. This comprehensive implementation ensures the efficient and automated management of indoor fish farming while providing users with real-time insights and control through the mobile application.